apply-pullreq: Ignore protocol and trailing .git when finding remote name
When looking through the list of known remotes for one that matches the
URL specified on the command line, ignore differences in protocol and
whether there is a trailing .git specified or not. This avoids the need
to manually fiddle with the URL string just because the submitter changed
their flow to use a different git protocol.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>