apply-pullreq: Use git commit --amend --signoff --no-edit
When modifying the merge commit to add the signoff line,
use git commit's options to do that without invoking an
editor (which requires git 1.7.9 or better). This is much
less ugly than the weird hack where we piped the old
git message into git commit's stdin, and also allows us
to drop the hard-coded line specifying the signoff to use.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>